requirements.txt
face_recognition
numpy
opencv-python
PyYAML
requests